feat: Implement user management features
- Add user fetching with pagination and search functionality
- Implement user deletion with confirmation dialog
- Enhance user interface for user management
- Update API paths for user-related operations
- Refactor pagination logic to use int64 for consistency

+++ server/internal/api/category.go
@@ -285,7 +285,7 @@ func (sv *Server) getAdminCategoriesHandler(c *gin.Context) {
 			Page:            query.Page,
 			PageSize:        query.PageSize,
 			Total:           count,
-			TotalPages:      int(count / int64(query.PageSize)),
+			TotalPages:      count / int64(query.PageSize),
 			HasNextPage:     count > int64(query.Page*query.PageSize),
 			HasPreviousPage: query.Page > 1,
 		}, nil,
